% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-decay.R
\name{decay}
\alias{decay}
\title{Decay}
\usage{
decay(x, half_lifes)
}
\arguments{
\item{x}{Numeric vector to decay.}

\item{half_lifes}{(Numeric scalar) The number of half lifes that have elapsed.}
}
\value{
A numeric vector.
}
\description{
Takes a numeric vector and applies exponential decay.
}
