% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-decay.R
\name{spectral_decay}
\alias{spectral_decay}
\alias{spectral_decay.vec}
\alias{spectral_decay.coded_vec}
\alias{spectral_decay.vec_milne_pc_spectrum}
\title{Spectral decay}
\usage{
spectral_decay(x, half_life = 3, offset = FALSE, ...)

\method{spectral_decay}{vec}(x, half_life = 3, offset = FALSE, ...)

\method{spectral_decay}{coded_vec}(x, ...)

\method{spectral_decay}{vec_milne_pc_spectrum}(x, half_life = 3,
  offset = FALSE, ...)
}
\arguments{
\item{x}{Chord sequence as created by \code{\link[hrep]{vec}}.
This chord sequence will be coerced to a
\code{\link[hrep]{milne_pc_spectrum}} representation.}

\item{half_life}{(Numeric scalar)
Half-life of the exponential decay, in units of chords.}

\item{offset}{(Logical scalar)
If \code{FALSE}, element i of the output vector
corresponds to the accumulated weight
at the timepoint immediately after the onset of chord i.
If \code{TRUE}, element i corresponds to
the accumulated weight at the end of chord i
(i.e. just before chord i + 1).}

\item{...}{Arguments passed to \code{\link[hrep]{milne_pc_spectrum}}.}
}
\value{
A vector (\code{\link[hrep]{vec}}) of Milne pitch-class spectra
(\code{\link[hrep]{milne_pc_spectrum}})
produced by accumulating pitch-class spectra over time
with exponential decay.
}
\description{
Given a chord sequence, this function simulates the accumulation
of acoustic information in sensory memory.
}
\details{
Chords are represented as objects of class
\code{\link[hrep]{milne_pc_spectrum}}.
The weights of these pitch-class spectra decay
exponentially over time.
}
