% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-similarity.R
\name{spectral_similarity}
\alias{spectral_similarity}
\alias{spectral_similarity.coded_vec}
\alias{spectral_similarity.vec}
\title{Spectral similarity}
\usage{
spectral_similarity(x, half_life = 3, ...)

\method{spectral_similarity}{coded_vec}(x, ...)

\method{spectral_similarity}{vec}(x, half_life = 3, ...)
}
\arguments{
\item{x}{Chord sequence as created by \code{\link[hrep]{vec}}.
This chord sequence will be coerced to a
\code{\link[hrep]{milne_pc_spectrum}} representation.}

\item{half_life}{(Numeric scalar)
Half-life of the exponential decay, in units of chords.}

\item{...}{Arguments passed to \code{\link[hrep]{milne_pc_spectrum}}.}
}
\value{
Numeric vector, where the ith element corresponds to the spectral similarity
of chord i to its preceding acoustic context.
The first element is always NA.
}
\description{
Computes the spectral similarity of successive chords in a chord sequence
to an integrated acoustic image of their recent context,
as computed by \code{\link{spectral_decay}}.
}
